Move button logic from lin.py to a separate class
